Add Reference type. Consolidate and add documentation links to types.
