--- src/fhirtypes/Extension.ts
@@ -1,10 +1,0 @@
-/**
- * Represents the FHIR R4 Extension element.
- * The FHIR type's `value[x]` field can have many names and types,
- * as listed on the accompanying page describing the FHIR type.
- *
- * @see {@link http://hl7.org/fhir/R4/extensibility.html#Extension}
- */
-export type Extension = {
-  url: string;
-};
