+++ src/fhirtypes/index.ts
@@ -3,6 +3,4 @@ export * from './metaDataTypes';
 export * from './primitiveTypes';
 export * from './specialTypes';
 export * from './ElementDefinition';
-export * from './Extension';
-export * from './Resource';
 export * from './StructureDefinition';
